backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
python-multipart==0.0.6
pydantic==1.10.12
PyJWT==2.8.0
email-validator==2.3.0
pytest==7.4.0
requests==2.31.0
transformers==4.41.2
torch
SQLAlchemy==2.0.20
passlib[bcrypt]==1.7.4
python-dotenv==1.2.1
dnspython==2.4.2
pyspf==2.0.14
dkimpy==1.1.5
python-whois==0.9.3
beautifulsoup4==4.12.2
lxml==4.9.3
whois==0.9.27
certifi
